Use separate layer for opacity.
