+++ assets/src/edit-story/elements/square/index.js
@@ -39,6 +39,7 @@ export const resizeRules = {
 
 export const panels = [
   PanelTypes.SIZE_POSITION,
+  PanelTypes.LAYER_STYLE,
   PanelTypes.STYLE,
   PanelTypes.LINK,
   PanelTypes.FILL,
