+++ assets/src/edit-story/elements/text/index.js
@@ -51,6 +51,7 @@ export const resizeRules = {
 
 export const panels = [
   PanelTypes.SIZE_POSITION,
+  PanelTypes.LAYER_STYLE,
   PanelTypes.TEXT_STYLE,
   PanelTypes.LINK,
 ];
